requirements.txt
# ChurnMaster Pro — Dependencies (Vercel-ready)
flask>=2.3.0
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
openpyxl>=3.1.0
Werkzeug>=3.0.0
python-dotenv>=1.0.0
# gunicorn is NOT needed on Vercel — Vercel provides its own WSGI server
# waitress>=3.0.0     # optional: Windows local server
# xgboost>=2.0.0      # optional: if your model_bundle uses XGBoost
# imbalanced-learn>=0.11.0  # optional: if you retrain with ChurnIQ Pro train.py
